requirements.txt
opencv-python
streamlit
numpy
paho-mqtt
mediapipe